Calcクラス分け、iPhone11黒画面調査中。
